requirements.txt
streamlit>=1.28.0
flask
PyPDF2>=3.0.0
python-pptx>=0.6.21
python-docx>=0.8.11
nltk>=3.8
requests>=2.31.0
